Refactor background logic based on messages
