--- packages/sanes-chrome-extension/src/extension/background/internalHandler/requestMeta.ts
@@ -1,16 +1,0 @@
-/**
- * All the information we want to pass with an incoming JSON-RPC request
- * into the authorization callbacks.
- */
-export interface RequestMeta {
-  readonly senderUrl: string;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isRequestMeta(data: unknown): data is RequestMeta {
-  if (typeof data !== 'object' || data === null) {
-    return false;
-  }
-
-  return typeof (data as RequestMeta).senderUrl === 'string';
-}
